Extract immutables style into @ImmutableStyle
This will also silence a compile type warning for packages that
depend on tablier and doesn't build with the immutables jar in
the classpath. The warning:

    Unknown enum constant ImplementationVisibility.PACKAGE

Is harmless in this case, but happens because the annotation
is retained on a class level, so when the class is imported for
the first time the compiler detects that the symbol doesn't
exist and whines.