Make sure MARKDOWN_OUTPUT_PATH exists
Releases use a fresh checkout in a separate subtree, meaning
that a folder that doesn't exist in the git repository won't
exist during the build (in the `release:perform` context).

This patch just adds the equivalent of `mkdir -p` to guarantee
that the directory does exist before generating the files.