Add some metadata for iOS users
Requests for `apple-touch-icon` are the top 404 error I've been
getting and I guess that's a sign of use? Can't think of a robot
fetching these.

Regardless, if it they are real users I'd like to provide a better
"add to home-screen" experience and this additional metadata does
the trick really well.